# docs/swagger.yaml
basePath: /api
definitions:
  handlers.UpdateUserRolesRequest:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
...
  models.Todo:
    properties:
      completed:
        type: boolean
      created_at:
        description: CreatedAt CustomTime `json:"created_at"`
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        description: UpdatedAt CustomTime `json:"updated_at"`
        type: string
    required:
    - title
    type: object
...
  response.PagedList:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
  response.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: Business status code
        type: integer
      data:
        description: Response data
      message:
        description: Response message
        type: string
      trace_id:
        description: Trace ID for request tracking
        type: string
    type: object
...
  services.CreateTodoRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  services.UpdateTodoRequest:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  v1.UpdateMenuRolesRequest:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A modern Go admin scaffold API server.
  title: Go Admin Scaffold API
  version: "1.0"
paths:
...
  /admin/v1/menus/{id}/roles:
    put:
      consumes:
      - application/json
      description: Update roles associated with a menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role IDs
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateMenuRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update menu roles
      tags:
      - menus
...
  /admin/v1/todos:
    get:
      consumes:
      - application/json
      description: Get a paginated list of todos
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PagedList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo info
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/services.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Todo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create todo
      tags:
      - todos
  /admin/v1/todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete todo
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: Get todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Todo'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get todo
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo info
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/services.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Todo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update todo
      tags:
      - todos
...
  /admin/v1/users/{id}/roles:
    put:
      consumes:
      - application/json
      description: Updates the roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role IDs
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update user roles
      tags:
      - User
...
swagger: "2.0"